+++ io.py
@@ -77,14 +77,17 @@ def saveResult(blockID, resp, columns=['respKey','RT'], stim=None, stim_columns=
 
 # send trigger
 def sendTrigger(data, mode='P'):
-    if mode=='P':
-        shared.Objdll.Out32(shared.setting['port'],0)
-    elif mode=='S': 
-        shared.ser.write(bytes(data,encoding='utf-8')) # send a string which might change
-        # shared.ser.write(b'something') # send a string directly
+    try:
+        if mode=='P':
+            shared.Objdll.Out32(shared.setting['port'],0)
+        elif mode=='S': 
+            shared.ser.write(bytes(data,encoding='utf-8')) # send a string which might change
+            # shared.ser.write(b'something') # send a string directly
 
-        # n=int('0b00010001',2)
-        # shared.ser.write(n.to_bytes((n.bit_length()+7)//8, 'big')) # send a binary code
-    else:
-        raise ValueError('Only support "S" or "P" (serial/parallel) mode!')
+            # n=int('0b00010001',2)
+            # shared.ser.write(n.to_bytes((n.bit_length()+7)//8, 'big')) # send a binary code
+        else:
+            raise ValueError('Only support "S" or "P" (serial/parallel) mode!')
+    except:
+        print('The port might be closed.')