+++ test/EXAMPLE -- Visual Experiment.py
@@ -7,14 +7,11 @@ sys.path.append('../../')
 from expy import * # Import the needed functions
 start() # Initiate the experiment environment
 
-# name = getInput('Please enter your name:')
-instruction(shared.setting['instruction1'])
-
 def trial(word,pos):
     drawText(word,x=pos) # Draw text on the canvas
     show() # Display current canvas
 
-    key = waitForEvent({K_f:'Left',K_j:'Right'}) # Waiting for pressing 'F' or 'J'
+    key,rt = waitForResponse({K_f:'Left',K_j:'Right'}) # Waiting for pressing 'F' or 'J'
     if key==word:
         alertAndGo('Correct!',1000) # Display something in 1s
     else:
@@ -22,6 +19,8 @@ def trial(word,pos):
 
     show(500) # Pause (show a screen during 500ms)
 
+# name = getInput('Please enter your name:')
+instruction(shared.setting['instruction1'])
 alertAndGo('The experiment will start after 3s.') # Display something in 3s(default)
 
 stimuli = [('Left',-0.5),
@@ -30,6 +29,7 @@ stimuli = [('Left',-0.5),
             ('Left',0.5),
             ('Left',-0.5),
             ('Right',-0.5)]
+
 for word,pos in stimuli:
     trial(word,pos) # Call the trial function with different parameters